Move TupleHash to utilities
